Move topic-level-assign-menu to components/
