+++ assets/javascripts/discourse/components/topic-level-assign-menu.js
